requirements.txt

streamlit
langgraph
langchain
langchain-core
langchain-community
langchain-groq
duckduckgo-search
python-dotenv
requests
supabase
faiss-cpu
chromadb
sentence-transformers
